Moving MvxPropertyChangingEventArgs to separate file
